pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = ["Sergey Shevchenko <sergiishevchenko017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
black = "^24.10.0"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
pydantic-settings = "^2.6.1"
slowapi = "^0.1.9"
passlib = "^1.7.4"
fastapi-mail = "^1.4.2"
cloudinary = "^1.41.0"
libgravatar = "^1.0.4"
jwt = "^1.3.1"
python-jose = "^3.3.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
aiocache = "^0.12.3"
aioredis = "^2.0.1"
aiosqlite = "^0.20.0"
sphinx = "^8.1.3"
bcrypt = "^4.2.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
